validate_history

+++ src/searchistory/src/models/HistoryModel.ts
@@ -9,6 +9,9 @@ import {
 } from 'firebase/firestore';
 
 import { db } from "../firebase/config";
+import { sanitize, reSanitize } from "../composable/sanitize";
+import { historyVali } from "../composable/validate";
+import cutWord from "../composable/cutWord";
 
 type HistoryStatus = 'all' | 'pending' | 'unsolved' | 'solved'
 
@@ -51,7 +54,7 @@ class HistoryModel extends PostCoreModel {
         this.topicDocID = topicObj.topicDocID;
     }
   }
-  // 登録
+  //============= 登録 =============
   static async register(
     url: string,
     title: string,
@@ -60,24 +63,41 @@ class HistoryModel extends PostCoreModel {
     files: FileInfo[],
     topicDocID: string,
   ) {
+    if (title.length === 0) {
+      if (!confirm("タイトルが空欄ですがよろしいですか?")) return;
+    }
+    // バリデーション
+    const valiResult = historyVali(url, title, content);
+    if (valiResult !== "") {
+      alert(valiResult);
+      return false;
+    }
+
     const { existFiles, deleteFiles } = super.splitFiles(files, content);
     super.deleteImgFromStorage(deleteFiles);
     const newHistoryRef = doc(collection(db, 'topic', topicDocID, 'history'));
 
-    await setDoc(newHistoryRef, {
-      url,
-      title,
-      content,
-      uid,
-      status: HISTORY_STATUS.PENDING,
-      createdAt: serverTimestamp(),
-      updatedAt: serverTimestamp(),
-      docID: newHistoryRef.id,
-      files: existFiles,
-      topicDocID
-    });
+    try {
+      await setDoc(newHistoryRef, {
+        url,
+        title,
+        content: sanitize(content),
+        uid,
+        status: HISTORY_STATUS.PENDING,
+        createdAt: serverTimestamp(),
+        updatedAt: serverTimestamp(),
+        docID: newHistoryRef.id,
+        files: existFiles,
+        topicDocID
+      });
+      return true;
+    } catch (e) {
+      console.log(e);
+      alert("エラーが発生しました。");
+      return false;
+    }
   }
-  // 更新
+  //============= 更新 =============
   static async update(
     url: string,
     title: string,
@@ -88,26 +108,42 @@ class HistoryModel extends PostCoreModel {
     topicDocID: string,
     status: HistoryStatus
   ) {
-    console.log("⬇︎【ログ】", "");
+    if (title.length === 0) {
+      if (!confirm("タイトルが空欄ですがよろしいですか?")) return;
+    }
+    // バリデーション
+    const valiResult = historyVali(url, title, content);
+    if (valiResult !== "") {
+      alert(valiResult);
+      return false;
+    }
 
-    const margeFiles = [...targetHistoryFiles, ...files]
-    const { existFiles, deleteFiles } = super.splitFiles(margeFiles, content);
-    super.deleteImgFromStorage(deleteFiles);
-    const updateHistoryRef = doc(db, 'topic', topicDocID, 'history', docID);
-    await updateDoc(updateHistoryRef, {
-      url,
-      title,
-      content,
-      files: existFiles,
-      updatedAt: serverTimestamp(),
-      status
-    });
+    try {
+      const margeFiles = [...targetHistoryFiles, ...files]
+      const { existFiles, deleteFiles } = super.splitFiles(margeFiles, content);
+      super.deleteImgFromStorage(deleteFiles);
+      const updateHistoryRef = doc(db, 'topic', topicDocID, 'history', docID);
+
+      await updateDoc(updateHistoryRef, {
+        url,
+        title,
+        content: sanitize(content),
+        files: existFiles,
+        updatedAt: serverTimestamp(),
+        status
+      });
+      return true;
+    } catch (e) {
+      console.log(e);
+      alert("エラーが発生しました。");
+      return false;
+    }
   }
-  // 削除
+  //============= 削除 =============
   async delete(topicDocID: string) {
+    if (!confirm(`この調査履歴を削除しますか?`)) return;
     const updateHistoryRef = doc(db, 'topic', topicDocID, 'history', this.docID);
-    await deleteDoc(updateHistoryRef)
-
+    await deleteDoc(updateHistoryRef);
   };
 }
 export { HistoryModel, HistoryStatus, HISTORY_STATUS, HISTORY_STATUS_WORD };
